(PC-34026) refacto(HorizontalTile): use children props
