+++ src/ui/components/tiles/HorizontalTile.tsx
@@ -1,89 +1,33 @@
 import React, { FC, PropsWithChildren } from 'react'
-import { FlexStyle } from 'react-native'
+import { StyleProp, ViewStyle } from 'react-native'
 import styled from 'styled-components/native'
 
 import { CategoryIdEnum } from 'api/gen'
-import { Tag } from 'ui/components/Tag/Tag'
 import { OfferImage } from 'ui/components/tiles/OfferImage'
-import { OfferName } from 'ui/components/tiles/OfferName'
-import { RightFilled } from 'ui/svg/icons/RightFilled'
 import { getSpacing } from 'ui/theme'
-import { TypoDS } from 'ui/theme/designSystemTypographie'
 
 export type HorizontalTileProps = PropsWithChildren<{
-  title?: string
   categoryId: CategoryIdEnum
   imageUrl?: string
-  distanceToOffer?: string
-  price?: string
-  withRightArrow?: boolean
+  style?: StyleProp<ViewStyle>
 }>
 
 export const HorizontalTile: FC<HorizontalTileProps> = ({
-  title = '',
   categoryId,
   imageUrl,
-  distanceToOffer,
-  price,
-  withRightArrow,
   children,
+  style,
 }) => {
   return (
-    <Container>
+    <Container style={style}>
       <OfferImage imageUrl={imageUrl} categoryId={categoryId} />
-      <Row flex={1} gap={getSpacing(4)} alignItems="space-between">
-        <Column flex={1}>
-          {distanceToOffer ? (
-            <OfferName title={title} />
-          ) : (
-            <Row>
-              <OfferNameContainer>
-                <OfferName title={title} />
-              </OfferNameContainer>
-              {withRightArrow ? <RightIcon testID="RightFilled" /> : null}
-            </Row>
-          )}
-          {children}
-          {price ? <TypoDS.BodyAccentS>{price}</TypoDS.BodyAccentS> : null}
-        </Column>
-        {distanceToOffer ? <DistanceTag label={`à ${distanceToOffer}`} /> : null}
-      </Row>
+      {children}
     </Container>
   )
 }
 
-const Flex = styled.View<FlexStyle>(({ flex, justifyContent, alignItems, gap, flexDirection }) => ({
-  flexDirection,
-  flex,
-  alignItems,
-  justifyContent,
-  gap,
-}))
-
-const Column = Flex
-
-const Row = styled(Flex).attrs({
+const Container = styled.View({
+  alignItems: 'center',
   flexDirection: 'row',
-  alignItems: 'center',
-})``
-
-const Container = styled(Row).attrs({
-  alignItems: 'center',
-  flex: 1,
   gap: getSpacing(4),
-})``
-
-const DistanceTag = styled(Tag)(({ theme }) => ({
-  backgroundColor: theme.colors.greyLight,
-}))
-
-const RightIcon = styled(RightFilled).attrs(({ theme }) => ({
-  size: theme.icons.sizes.extraSmall,
-}))({
-  flexShrink: 0,
-  marginLeft: getSpacing(1),
-})
-
-const OfferNameContainer = styled.View({
-  flexShrink: 1,
 })
